arreglo del delete users

+++ app/manage_users.py
@@ -1,4 +1,5 @@
 from app.models import User, User_Light_State,User_Temperature_State,Scheduled_events
+from app.server import delete_scheduled_event
 from app import db
 
 def create_user_full(form):
@@ -20,12 +21,22 @@ def delete_user_full(user):
     delete_user = User.query.filter_by(username=user).first()
     light_state_to_del= User_Light_State.query.filter_by(user=user).first()   
     temp_state_to_del= User_Temperature_State.query.filter_by(user=user).first()  
-    shceduled_events_to_del=Scheduled_events.quey.filter_by(user=user)
+    shceduled_events_to_del=Scheduled_events.query.filter_by(user=user)
     
+    #Agarro tods los pids del usuario que voy a eliminar y se los mando al metodo para que la libreria apscheduler se haga cargo
+    
+    pids=[]
+    for pid in shceduled_events_to_del:
+        pids.append(pid.pid)
+    
+    if len(pids)!=0:
+        delete_scheduled_event(pids)
+        db.session.delete(shceduled_events_to_del)
+
     db.session.delete(delete_user)
     db.session.delete(light_state_to_del)
     db.session.delete(temp_state_to_del)
-    db.session.delete(shceduled_events_to_del)
+    
     db.session.commit()
 
     return 'Congratulations, you have just deleted '+delete_user.username+ ' from the users list!'
